requirements-dev.txt
alembic
pydantic
python-dotenv
psycopg2
sqlalchemy-json
requests
fastapi
FastAPI-SQLAlchemy
uvicorn
gunicorn
aiofiles==0.6.0
python-multipart==0.0.5
jinja2==2.11.2
uvloop
httptools
pydantic[email]
email-validator==1.1.2
emails==0.6
pyjwt
python-jose==3.2.0
